Add v5 API module and update test setup
