+++ app/src/api.rs
@@ -1,4 +1,4 @@
-pub mod get;
+pub mod v5;
 mod request;
 
 use crate::constants::{
