+++ app/src/api/v5/get/market/get_historical_volatility.rs
@@ -1,7 +1,7 @@
 use crate::{
     api::{
         BybitApi,
-        get::Get,
+        v5::get::Get,
     },
     utils::{
         deserialize_f64,
@@ -36,7 +36,7 @@ impl BybitApi {
     ///
     /// ```rust
     /// use rsbit::api::{
-    ///     get::market::get_historical_volatility::{
+    ///     v5::get::market::get_historical_volatility::{
     ///         GetHistoricalVolatilityParameters,
     ///         GetHistoricalVolatilityCategory
     ///     },
