+++ app/src/api/v5/get/market/get_instruments_info.rs
@@ -12,7 +12,7 @@ use self::{
 use crate::{
     api::{
         BybitApi,
-        get::Get,
+        v5::get::Get,
     },
     constants::{
         CATEGORY_INVERSE,
@@ -44,7 +44,7 @@ impl BybitApi {
     ///
     /// ```rust
     /// use rsbit::api::{
-    ///     get::market::get_instruments_info::{
+    ///     v5::get::market::get_instruments_info::{
     ///         GetInstrumentsInfoParameters,
     ///         GetInstrumentsInfoCategory,
     ///         InstrumentsInfoResult,
