+++ app/src/api/v5/get/market/get_instruments_info/spot.rs
