+++ app/src/api/v5/get/market/get_public_recent_trading_history.rs
@@ -1,7 +1,7 @@
 use crate::{
     api::{
         BybitApi,
-        get::Get,
+        v5::get::Get,
     },
     utils::{
         deserialize_f64,
@@ -29,7 +29,7 @@ impl BybitApi {
     ///
     /// ```rust
     /// use rsbit::api::{
-    ///     get::market::get_public_recent_trading_history::{
+    ///     v5::get::market::get_public_recent_trading_history::{
     ///         GetPublicRecentTradingHistoryParameters,
     ///         GetPublicRecentTradingHistoryCategory
     ///     },
