+++ app/src/api/v5/get/market/get_tickers/linear.rs
