+++ app/tests/get/get_mark_price_kline_test.rs
@@ -1,12 +1,12 @@
-use rsbit::api::get::market::get_mark_price_kline::{
+use rsbit::api::v5::get::market::get_mark_price_kline::{
     GetMarkPriceKlineParameters,
     GetMarkPriceKlineCategory,
 };
-use crate::common::setup_api;
+use crate::common::setup_api_public;
 
 #[tokio::test]
 async fn test_get_mark_price_kline_success() {
-    let api = setup_api();
+    let api = setup_api_public();
     let categories = vec![
         (GetMarkPriceKlineCategory::Linear, "BTCUSDT".to_string(), "1".to_string()),
         (GetMarkPriceKlineCategory::Inverse, "BTCUSD".to_string(), "360".to_string()),
@@ -37,7 +37,7 @@ async fn test_get_mark_price_kline_success() {
 
 #[tokio::test]
 async fn test_get_mark_price_kline_fail() {
-    let api = setup_api();
+    let api = setup_api_public();
     let params = GetMarkPriceKlineParameters::new(
         GetMarkPriceKlineCategory::Linear,
         "XXXXXXX".to_string(),
